backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.7.2
attrs==24.2.0
axios==0.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
Django==4.2.8
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.27.2
exceptiongroup==1.2.2
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.0.2
openai==1.55.0
packaging==24.2
pandas==2.2.3
pd==0.0.4
pillow==11.0.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.21.0
